# Conflicts:
#	mobile/android/fenix/app/src/main/res/drawable-xxhdpi/firefox_sport.webp
#	mobile/android/fenix/app/src/main/res/values-it/strings.xml
#	mobile/android/fenix/app/src/main/res/values/ids.xml
#	mobile/android/fenix/app/src/main/res/values/static_strings.xml

+++ netwerk/protocol/http/HttpTransactionParent.cpp
@@ -150,12 +150,11 @@ nsresult HttpTransactionParent::Init(
   // TODO: Figure out if we have to implement nsIThreadRetargetableRequest in
   // bug 1544378.
   if (!SendInit(caps, infoArgs, *requestHead, ipcStream, requestContentLength,
-                requestBodyHasHeaders, browserId,
-                static_cast<uint8_t>(trafficCategory), requestContextID,
-                classOfService, initialRwin, responseTimeoutEnabled, mChannelId,
-                !!mTransactionObserver, throttleQueue, mIsDocumentLoad,
-                aParentIpAddressSpace, aLnaPermissionStatus, mRedirectStart,
-                mRedirectEnd)) {
+                requestBodyHasHeaders, browserId, trafficCategory,
+                requestContextID, classOfService, initialRwin,
+                responseTimeoutEnabled, mChannelId, !!mTransactionObserver,
+                throttleQueue, mIsDocumentLoad, aParentIpAddressSpace,
+                aLnaPermissionStatus, mRedirectStart, mRedirectEnd)) {
     return NS_ERROR_FAILURE;
   }
